requirements.txt
fastapi
uvicorn
httpx
python-dotenv
gunicorn
ollama